Removed duplicate React project and merged changes Installed Firebase dependency Updated readme
