--- netlify.toml
@@ -1,21 +1,0 @@
-# example netlify.toml
-[build]
-  command = "CI= npm run build"
-  functions = "netlify/functions"
-  publish = ""
-
-  ## Uncomment to use this redirect for Single Page Applications like create-react-app.
-  ## Not needed for static site generators.
-[[redirects]]
-  from = "/*"
-  to = "/index.html"
-  status = 200
-
-  ## (optional) Settings for Netlify Dev
-  ## https://github.com/netlify/cli/blob/main/docs/netlify-dev.md#project-detection
-  #[dev]
-  #  command = "yarn start" # Command to start your dev server
-  #  port = 3000 # Port that the dev server will be listening on
-  #  publish = "dist" # Folder with the static content for _redirect file
-
-  ## more info on configuring this file: https://docs.netlify.com/configure-builds/file-based-configuration/
